backend/requirements.txt
# DocuShrink AI - Backend Requirements
# =====================================

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# PDF Processing
PyPDF2==3.0.1

# Image Processing
Pillow==10.1.0

# OCR
pytesseract==0.3.10
pdf2image==1.16.3

# AI/ML (Optional - for HuggingFace models)
# Uncomment below for AI model support:
# transformers==4.35.0
# torch==2.1.0
# sentencepiece==0.1.99

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
